Fix a type bug that could cause incorrect material names in output file. This was introduced in r603 (July 7); anyone who has been using mcnp2cad to prepare DAGMC inputs in the last month should verify the correctness of their material group names.
Thanks to Brandon for catching this.

+++ mcnp2cad.cpp
@@ -135,7 +135,7 @@ class GeometryContext {
 
 protected:
 
-  std::string materialName( int mat, int rho ){
+  std::string materialName( int mat, double rho ){
     std::string ret;
     std::stringstream formatter;
     formatter << "mat_" << mat << "_rho_" << rho;
